+++ src/pytesmo/time_series/anomaly.py
@@ -22,32 +22,32 @@ def calc_anomaly(Ser,
 
     Parameters
     ----------
-    Ser : pandas.Series (index must be a DateTimeIndex)
-
-    window_size : float, optional
+    Ser : pandas.Series
+        Input data (index must be a DateTimeIndex)
+    window_size : float, optional (default: 35)
         The window-size [days] of the moving-average window to calculate the
         anomaly reference (only used if climatology is not provided)
-        Default: 35 (days)
-
-    climatology : pandas.Series (index: 1-366), optional
+    climatology : pandas.Series (index: 1-366), optional (default: None)
         if provided, anomalies will be based on the climatology
-
     timespan : [timespan_from, timespan_to], datetime.datetime(y,m,d), optional
         If set, only a subset
-
-    respect_leap_years : boolean, optional
-        If set then leap years will be respected during matching of the climatology
-        to the time series
-
-    return_clim : boolean, optional
+    respect_leap_years : boolean, optional (default: True)
+        If set then leap years will be respected during matching of the
+        climatology to the time series
+    return_clim : boolean, optional (default: False)
         if set to true the return argument will be a DataFrame which
         also contains the climatology time series.
         Only has an effect if climatology is used.
 
     Returns
     -------
-    anomaly : pandas.Series
-        Series containing the calculated anomalies
+    anomaly : pandas.Series or pandas.DataFrame
+        Series containing the calculated anomalies.
+        If `return_clim` is set to true, a DataFrame will be returned, where
+        one column contains the anomalies and another the climatology
+        broadcasted over the whole index. If a climatology with a 'std' column
+        was passed initially, this column will also be returned in the
+        DataFrame if `return_clim` is chosen.
     '''
 
     if climatology is not None:
@@ -70,7 +70,10 @@ def calc_anomaly(Ser,
         df['absolute'] = Ser
         df['doy'] = doys
 
-        clim = pd.DataFrame({'climatology': climatology})
+        if isinstance(climatology, pd.Series):
+            clim = pd.DataFrame({'climatology': climatology})
+        else:
+            clim = climatology
 
         df = df.join(clim, on='doy', how='left')
 
@@ -80,6 +83,8 @@ def calc_anomaly(Ser,
         if return_clim:
             anomaly = pd.DataFrame({'anomaly': anomaly})
             anomaly['climatology'] = df['climatology']
+            if 'std' in df.columns:
+                anomaly['climatology_std'] = df['std']
 
     else:
         reference = moving_average(Ser, window_size=window_size)
@@ -111,6 +116,7 @@ def calc_climatology(Ser,
                      moving_avg_orig=5,
                      moving_avg_clim=None,
                      median=False,
+                     std=False,
                      timespan=None,
                      fill=np.nan,
                      wraparound=True,
@@ -125,69 +131,62 @@ def calc_climatology(Ser,
 
     Parameters
     ----------
-    Ser : pandas.Series (index must be a DateTimeIndex or julian date)
-
-    moving_avg_orig : float, optional
-        The size of the moving_average window [days] that will be applied on the
-        input Series (gap filling, short-term rainfall correction)
-        Default: 5
-
-    moving_avg_clim : float, optional
-        The size of the moving_average window in days that will be applied on the
-        calculated climatology (long-term event correction). If None is passed, it
-        will be calculated from the 'output_freq' value:
+    Ser : pandas.Series
+        Time series to compute climatology for (index must be a
+        DateTimeIndex or julian date)
+    moving_avg_orig : float, optional (default: 5)
+        The size of the moving_average window [days] that will be applied on
+        the input Series (gap filling, short-term rainfall correction)
+    moving_avg_clim : float, optional (default: None)
+        The size of the moving_average window in days that will be applied on
+        the calculated climatology (long-term event correction).
+        If None is passed, it will be calculated from the 'output_freq' value:
             - 'day': 35
             - 'month': 3
-        Default: None
-
-    median : boolean, optional
+    median : boolean, optional (default: False)
         if set to True, the climatology will be based on the median conditions
-
+    std: boolean, optional (default: False)
+        if set to True, there will be 2 columns, one for the median or mean
+        and one of the standard deviation of the aggregated data points.
     timespan : [timespan_from, timespan_to], datetime.datetime(y,m,d), optional
         Set this to calculate the climatology based on a subset of the input
         Series
-
-    fill : float or int, optional
+    fill : float or int, optional (default: np.nan)
         Fill value to use for days on which no climatology exists
-
-    wraparound : boolean, optional
+    wraparound : boolean, optional (default: True)
         If set then the climatology is wrapped around at the edges before
         doing the second running average (long-term event correction)
-
-    respect_leap_years : boolean, optional
+    respect_leap_years : boolean, optional (default: False)
         If set then leap years will be respected during the calculation of
-        the climatology. Only valid with 'unit' value set to 'day'.
+        the climatology. Only valid with 'output_freq' value set to 'day'.
         Default: False
-
-    interpolate_leapday: boolean, optional
-        <description>. Only valid with 'unit' value set to 'day'.
+    interpolate_leapday: boolean, optional (default: False)
+        <description>. Only valid with 'output_freq' value set to 'day'.
         Default: False
-
-    fillna: boolean, optional
+    fillna: boolean, optional (default: True)
         If set, then the moving average used for the calculation of the
         climatology will be filled at the nan-values
-
-    min_obs_orig: int
+    min_obs_orig: int (default: 1)
         Minimum observations required to give a valid output in the first
         moving average applied on the input series
-
-    min_obs_clim: int
+    min_obs_clim: int (default: 1)
         Minimum observations required to give a valid output in the second
         moving average applied on the calculated climatology
-
-    output_freq: str, optional
+    output_freq: str, optional (default: 'day')
         Determines the output frequency (time unit) of the climatology
         calculation (independently of the 'Ser' input frequency).
         Currently, supported options are 'day', 'month'.
-        Default: 'day'
 
     Returns
     -------
-    climatology : pandas.Series
-        Containing the calculated climatology. The size of the series depends on
-        the type of climatology being calculated, based on the value of 'output_freq':
+    climatology : pandas.Series or pandas.DataFrame
+        Containing the calculated climatology. The size of the series depends
+        on the type of climatology being calculated, based on the value of
+        'output_freq':
             - 366 values for a daily climatology, behaving as a leap year
             - 12 values for a monthly climatology
+        If 'std' is set to True, the output will be a DataFrame with 2 columns:
+            'climatology' and 'std'.
     """
     # establish the moving window size
     default_moving_avg_clim = {"day": 35, "month": 3}
@@ -196,10 +195,10 @@ def calc_climatology(Ser,
         moving_avg_clim = default_moving_avg_clim[output_freq]
 
     if output_freq == "month" and moving_avg_clim > 5:
-        # in case someone changes unit but not moving_avg_clim a warning might be useful
-        warnings.warn(
-            f"Window for moving average of climatology set to {moving_avg_clim} months, is this intentional?"
-        )
+        # in case someone changes unit but not moving_avg_clim a warning
+        # might be useful
+        warnings.warn(f"Window for moving average of climatology set to "
+                      f"{moving_avg_clim} months, is this intentional?")
 
     if output_freq != "day":
         # irrelevant at lower frequencies than daily
@@ -231,10 +230,21 @@ def calc_climatology(Ser,
 
     if median:
         clim = Ser.groupby('unit').median()
+        clim.name = 'climatology'
     else:
         clim = Ser.groupby('unit').mean()
+        clim.name = 'climatology'
 
-    clim_ser = pd.Series(clim.values.flatten(), index=clim.index.values)
+    if std:
+        std_ser = Ser.groupby('unit').std()
+
+        clim_ser = pd.concat([
+            clim.loc[:, 0].rename(clim.name),
+            std_ser.loc[:, 0].rename('std')], axis=1)  # yapf: disable
+    else:
+        clim_ser = pd.DataFrame(
+            data={'climatology': clim.values.flatten()},
+            index=clim.index.values)
 
     clim_ser = clim_ser.reindex(np.arange(n_idx) + 1)
 
@@ -248,26 +258,27 @@ def calc_climatology(Ser,
         clim_ser.index = clim_ser.index + n_idx
         clim_ser = pd.concat([left_mirror, clim_ser, right_mirror])
 
-        clim_ser = moving_average(
-            clim_ser,
-            window_size=moving_avg_clim,
-            fillna=fillna,
-            min_obs=min_obs_clim)
+    clim_ser['climatology'] = moving_average(
+        clim_ser['climatology'],
+        window_size=moving_avg_clim,
+        fillna=fillna,
+        min_obs=min_obs_clim)
+
+    if wraparound:
         clim_ser = clim_ser.iloc[moving_avg_clim:-moving_avg_clim]
         clim_ser.index = index_old
-    else:
-        clim_ser = moving_average(
-            clim_ser,
-            window_size=moving_avg_clim,
-            fillna=fillna,
-            min_obs=min_obs_clim)
 
     # keep hardcoding as it's only for doys
     if interpolate_leapday and not respect_leap_years:
-        clim_ser[60] = np.mean((clim_ser[59], clim_ser[61]))
+        clim_ser.loc[60, :] = np.mean(
+            (clim_ser.loc[59, :], clim_ser.loc[61, :]))
     elif interpolate_leapday and respect_leap_years:
-        clim_ser[366] = np.mean((clim_ser[365], clim_ser[1]))
+        clim_ser.loc[366, :] = np.mean(
+            (clim_ser.loc[365, :], clim_ser.loc[1, :]))
 
     clim_ser = clim_ser.fillna(fill)
 
-    return clim_ser
+    if len(clim_ser.columns) == 1:
+        return clim_ser.iloc[:, 0]
+    else:
+        return clim_ser